64-bit fix for mfd from Robert Tsai.  Closes #32.
git-svn-id: http://svn.mythtv.org/svn/trunk@6771 7dbf422c-18fa-0310-86e9-fd20926502f2

+++ mfd/mfd/plugins/zeroconfig/apple/mDNSClientAPI.h
@@ -477,8 +477,8 @@ typedef   signed char  mDNSs8;
 typedef unsigned char  mDNSu8;
 typedef   signed short mDNSs16;
 typedef unsigned short mDNSu16;
-typedef   signed long  mDNSs32;
-typedef unsigned long  mDNSu32;
+typedef   signed int   mDNSs32;
+typedef unsigned int   mDNSu32;
 
 // To enforce useful type checking, we make mDNSInterfaceID be a pointer to a dummy struct
 // This way, mDNSInterfaceIDs can be assigned, and compared with each other, but not with other types
